Add comment styling

+++ src/main/java/com/cm/contentmanagementapp/services/BlogPostServiceImpl.java
@@ -102,7 +102,8 @@ public class BlogPostServiceImpl implements BlogPostService {
     public boolean updateBlogImage(BlogPost blogPost, MultipartFile file) {
         Path filePath = Paths.get("uploads/blogs/blogImg");
 
-        if (!blogPost.getPost().getImage().getImgName().isEmpty()) {
+        if (blogPost.getPost().getImage().getImgName() != null
+                && !blogPost.getPost().getImage().getImgName().isEmpty()) {
             fileService.deleteIfExists(filePath, blogPost.getPost().getImage().getImgName());
         }
